Added Scott's code in, still in java. Fixed bugs. Modified equation for inv_distance map

+++ nameseparation/dimension.cpp
@@ -141,7 +141,7 @@ void Dimension::setNumOfBins(int num)
     numOfBins = num;
 }
 
-void Dimension::setMinMax(int min, int max)
+void Dimension::setMinMax(double min, double max)
 {
     minValue=min;
     maxValue=max;
